+++ tests/assets/import2.ts
@@ -1,0 +1,5 @@
+export type Foo = Buffer;
+
+export interface Bar {
+  buffer: Buffer;
+}