+++ tests/assets/type9.ts
@@ -1,0 +1,13 @@
+const keyValue = {
+  COMPONENT_ERROR__LNK_LABEL: '',
+  COMPONENT_ERROR__LNK_MESSAGE: '',
+  COMPONENT_ERROR__CNTEXT_MENU_LINK_LABEL: '',
+  COMPONENT_ERRR__CNTEXT_MENU_LINK_MESSAGE: '',
+  E: '',
+  D: '',
+  C: '',
+  B: '',
+  A: '',
+};
+
+export type KeyValue = typeof keyValue;